+++ ios/StripeSdk.swift
@@ -1116,13 +1116,17 @@ class StripeSdk: RCTEventEmitter, STPBankSelectionViewControllerDelegate, UIAdap
             resolve(Errors.createError(ErrorType.Failed, "You must provide `cardLastFour`"))
             return
         }
-        let (canAddCard, status) = PushProvisioningUtils.canAddCardToWallet(last4: last4,
-                                                 primaryAccountIdentifier: params["primaryAccountIdentifier"] as? String ?? "",
-                                                 testEnv: params["testEnv"] as? Bool ?? false)
-        resolve([
-            "canAddCard": canAddCard,
-            "details": ["status": status?.rawValue],
-        ])
+        PushProvisioningUtils.canAddCardToWallet(
+            last4: last4,
+            primaryAccountIdentifier: params["primaryAccountIdentifier"] as? String ?? "",
+            testEnv: params["testEnv"] as? Bool ?? false,
+            hasPairedAppleWatch: params["hasPairedAppleWatch"]  as? Bool ?? false)
+        { canAddCard, status in
+            resolve([
+                "canAddCard": canAddCard,
+                "details": ["status": status?.rawValue],
+            ])
+        }
     }
 
     @objc(isCardInWallet:resolver:rejecter:)
@@ -1135,7 +1139,7 @@ class StripeSdk: RCTEventEmitter, STPBankSelectionViewControllerDelegate, UIAdap
             resolve(Errors.createError(ErrorType.Failed, "You must provide `cardLastFour`"))
             return
         }
-        resolve(["isInWallet": PushProvisioningUtils.passExistsWith(last4: last4)])
+        resolve(["isInWallet": PushProvisioningUtils.getPassLocation(last4: last4) != nil])
     }
 
     @objc(collectBankAccountToken:resolver:rejecter:)