+++ ios/Tests/PushProvisioningTests.swift
@@ -11,19 +11,22 @@ import XCTest
 
 class PushProvisioningTests: XCTestCase {
     func testCanAddCardToWalletInTestMode() throws {
-        let (canAddCard, status) = PushProvisioningUtils.canAddCardToWallet(last4: "4242",
+        PushProvisioningUtils.canAddCardToWallet(last4: "4242",
                                                  primaryAccountIdentifier: "",
-                                                 testEnv: true)
-        XCTAssertEqual(canAddCard, true)
-        XCTAssertEqual(status, nil)
+                                                 testEnv: true, hasPairedAppleWatch: false) { canAddCard, status in
+            XCTAssertEqual(canAddCard, true)
+            XCTAssertEqual(status, nil)
+        }
     }
-
+    
     func testCanAddCardToWalletInLiveMode() throws {
-        let (canAddCard, status) = PushProvisioningUtils.canAddCardToWallet(last4: "4242",
+        PushProvisioningUtils.canAddCardToWallet(last4: "4242",
                                                  primaryAccountIdentifier: "",
-                                                 testEnv: false)
-        XCTAssertEqual(canAddCard, false)
-        XCTAssertEqual(status, PushProvisioningUtils.AddCardToWalletStatus.MISSING_CONFIGURATION)
+                                                 testEnv: false,
+                                                 hasPairedAppleWatch: false) { canAddCard, status in
+            XCTAssertEqual(canAddCard, false)
+            XCTAssertEqual(status, PushProvisioningUtils.AddCardToWalletStatus.MISSING_CONFIGURATION)
+        }
     }
     
     func testCanAddPaymentPassInTestMode() throws {
@@ -42,8 +45,8 @@ class PushProvisioningTests: XCTestCase {
     
     func testCheckIfPassExists() throws {
         XCTAssertEqual(
-            PushProvisioningUtils.passExistsWith(last4: "4242"),
-            false
+            PushProvisioningUtils.getPassLocation(last4: "4242"),
+            nil
         )
     }
 }